Update CommandUI.lua
Added a click to delete toggle this will be saved with click teleport and requires you to hold X when clicking to delete things.
Added player search which will make it much easier to find your intended target in all the drop down player lists specially in larger servers.
Both of these have been put in the server tab for the time being.

Currently looking for a new UI to provide more room and potentially less tab switching, may also make my own but that has been on the back burner for years. Other than all that just fixed some typos and what not.